# lib/api-zod/src/generated/types/recommendJobsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerForge AI API
 * OpenAPI spec version: 0.1.0
 */

export interface RecommendJobsRequest {
  skills: string[];
  interests: string[];
  targetRole?: string | null;
}