# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareerForge AI API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfile {
  id: number;
  userId: number;
  fullName: string;
  email: string;
  college?: string | null;
  branch?: string | null;
  year?: string | null;
  skills: string[];
  interests: string[];
  targetRole?: string | null;
  bio?: string | null;
  completionPercentage: number;
  updatedAt: Date;
}